fix: add new variable to specifically set the AppSet destination
The ApplicationSet should be deployed on the management cluster and not alongside the Application, so the destination for both resources is not actually the same. Because of that, a variable was introduced to specifically set the destination of the Application that deploys the ApplicationSets, although most of the times this variable will always be `in-cluster`.